+++ javaliproject/src/main/java/br/ufc/quixada/javaliproject/service/AlunoServiceImpl.java
@@ -30,7 +30,7 @@ public class AlunoServiceImpl implements AlunoService {
 	}
 
 	@Override
-	public void remover(Long id) {
+	public void remover(int id) {
 		Aluno aluno = alunoRepository.findById(id);
 		alunoRepository.remover(aluno);
 	}
